actually, wrong type for printing of sise_t

+++ usr.sbin/dhcp/common/bpf.c
@@ -198,8 +198,8 @@ void if_register_receive (info)
 	info->rbuf_max = sz;
 	info -> rbuf = malloc (info -> rbuf_max);
 	if (!info -> rbuf)
-		error ("Can't allocate %d bytes for bpf input buffer.",
-		    info -> rbuf_max);
+		error ("Can't allocate %lu bytes for bpf input buffer.",
+		    (unsigned long) info -> rbuf_max);
 	info -> rbuf_offset = 0;
 	info -> rbuf_len = 0;
